added comments and inheritance for messages
